pyproject.toml
[project]
name = "oz-main-be-12-team2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "django-allauth[socialaccount]>=65.0.2",
    "coverage>=7.10.6",
    "drf-spectacular",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
    "mypy>=1.17.1",
    "pytest>=8.0.0",
    "pytest-django>=4.0.0",
    "coverage>=7.10.6",
]

[tool.ruff]
line-length = 120
exclude = ["migrations", ".venv"]
select = ["E", "F", "I"]

[tool.ruff.lint]
# 자동수정은 임포트 정렬(I)만 허용
select = ["I"]
# 필요 시 보고 전용 규칙(F/E/W 등)은 실행 스크립트에서 --extend-select로 켜야함
fixable = ["ALL"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "--tb=short"
python_files = "tests.py test_*.py *_test.py" 

# test 코드 빼고 coverage 측정
[tool.coverage.run]
omit = ["*/test_*.py"]
